Use .test-queue over /tmp
Sometimes /tmp isn't writeable for the user running test-queue.
While an admittedly unusual, and maybe even rare occurrence, this
has proven troublesome in the past.

Defaulting to using .test-queue should remove any concerns/problems
on the topic of permissions, and provide an easy means to gitignore
this sockets/outputs/stats.